+++ geneticAlgorith/src/no/uib/mof077/shortbytes/neural/NodeLayer.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 
 public class NodeLayer {
 	public ArrayList<Node> nodeList = new ArrayList<>();
-	public float learningRate;
+	public double learningRate;
 	
-	public NodeLayer(ArrayList<Node> nodeList, float learningRate) {
+	public NodeLayer(ArrayList<Node> nodeList, double learningRate) {
 		this.nodeList = nodeList;
 		this.learningRate = learningRate;
 		for (Node node : nodeList) {
